REFACTOR strategies
