+++ avalanche/training/strategies/cumulative.py
@@ -1,0 +1,62 @@
+from typing import Optional, List
+
+from torch.nn import Module
+from torch.optim import Optimizer
+from torch.utils.data import ConcatDataset
+
+from avalanche.logging import default_logger
+from avalanche.training.plugins import StrategyPlugin, EvaluationPlugin
+from avalanche.training.strategies import BaseStrategy
+
+
+class Cumulative(BaseStrategy):
+
+    def __init__(self, model: Module, optimizer: Optimizer, criterion,
+                 train_mb_size: int = 1, train_epochs: int = 1,
+                 eval_mb_size: int = None, device=None,
+                 plugins: Optional[List[StrategyPlugin]] = None,
+                 evaluator: EvaluationPlugin = default_logger, eval_every=-1):
+        """ Cumulative strategy. At each experience,
+            train model with data from all previous experiences and current
+            experience.
+
+        :param model: The model.
+        :param optimizer: The optimizer to use.
+        :param criterion: The loss criterion to use.
+        :param train_mb_size: The train minibatch size. Defaults to 1.
+        :param train_epochs: The number of training epochs. Defaults to 1.
+        :param eval_mb_size: The eval minibatch size. Defaults to 1.
+        :param device: The device to use. Defaults to None (cpu).
+        :param plugins: Plugins to be added. Defaults to None.
+        :param evaluator: (optional) instance of EvaluationPlugin for logging
+            and metric computations.
+        :param eval_every: the frequency of the calls to `eval` inside the
+            training loop.
+                if -1: no evaluation during training.
+                if  0: calls `eval` after the final epoch of each training
+                    experience.
+                if >0: calls `eval` every `eval_every` epochs and at the end
+                    of all the epochs for a single experience.
+        """
+
+        super().__init__(
+            model, optimizer, criterion,
+            train_mb_size=train_mb_size, train_epochs=train_epochs,
+            eval_mb_size=eval_mb_size, device=device, plugins=plugins,
+            evaluator=evaluator, eval_every=eval_every)
+
+        self.dataset = {}  # cumulative dataset
+
+    def adapt_train_dataset(self, **kwargs):
+
+        super().adapt_train_dataset(**kwargs)
+
+        curr_task_id = self.experience.task_label
+        curr_data = self.experience.dataset
+        if curr_task_id in self.dataset:
+            cat_data = ConcatDataset([self.dataset[curr_task_id],
+                                      curr_data])
+            self.dataset[curr_task_id] = cat_data
+        else:
+            self.dataset[curr_task_id] = curr_data
+        self.adapted_dataset = self.dataset
